new_approach/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
colorful==0.5.8
distlib==0.4.0
dm-tree==0.1.9
Farama-Notifications==0.0.4
filelock==3.20.1
frozenlist==1.8.0
fsspec==2026.1.0
google-api-core==2.28.1
google-auth==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gymnasium==1.1.1
idna==3.11
importlib_metadata==8.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lz4==4.4.5
msgpack==1.1.2
multidict==6.7.0
numpy==2.4.1
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
ormsgpack==1.7.0
packaging==25.0
pandas==2.3.3
platformdirs==4.5.1
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
ray==2.52.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scipy==1.17.0
six==1.17.0
smart_open==7.5.0
tensorboardX==2.6.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
virtualenv==20.35.4
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
